+++ api/Interfaces/ITouristSpotRequestValidation.cs
@@ -4,7 +4,7 @@ namespace api.Interfaces
 {
     public interface ITouristSpotRequestValidator
     {
-        IActionResult ValidateLocationRequest(double latitude, double longitude, int? minPrice, int? maxPrice, double? minRating);
-        IActionResult ValidateNameRequest(string name, int? minPrice, int? maxPrice, double? minRating);
+        IActionResult ValidateLocationRequest(double latitude, double longitude, int? minPrice, int? maxPrice, decimal? minRating);
+        IActionResult ValidateNameRequest(string name, int? minPrice, int? maxPrice, decimal? minRating);
     }
 }
